+++ maskgen/ui/collapsing_frame.py
